Lazy load matrix-js-sdk when running as SPA
